pyproject.toml
[tool.poetry]
name = "orchidarium"
version = "0.0.1"
description = "Orchestration and scripts for publishing metrics from sensors in an orchidarium to InfluxDB for Grafana dashboards"
authors = ["Emma Doyle <emma.ann.doyle@gmail.com>"]
maintainers = ["Emma Doyle <emma.ann.doyle@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [
  { include = "orchidarium", from = "src" },
]
include = [
  "LICENSE"
]
keywords = [
    "python",
    "orchids",
    "plants",
    "terrarium"
]

[tool.poetry.dependencies]
python = "^3.10"
pyusb = "^1.3.1"
sht20 = "^0.1.4"
smbus2 = "^0.5.0"

flask = "^3.1.2"
bitstring = "^4.3.1"
influxdb3-python = "^0.16.0"
hid = "^1.0.8"
[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
mypy = "^1.18.2"
pylint = "^3.3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
orchidarium = "orchidarium.entrypoint:cli"

[tool.mypy]
python_version = "3.11"
strict_optional = "true"
ignore_missing_imports = "true"